Cleanup and small refactors across tauri src
Apply numerous small code cleanups and idiomatic Rust refactors across the tauri source:

- Use saturating_sub for brace-depth decrement in JSON scanner.
- Simplify string newline replacements with slice overloads and split helpers.
- Use range pattern (..=) for weather codes; use find with slices for path parsing.
- Remove redundant returns/drops and simplify conditional logic (including early returns and boolean expressions).
- Apply Option/Question-mark idiom and simplify iterator logic where possible.
- Add clippy allowances/comments where appropriate (too_many_arguments, type_complexity).
- Replace construction of mutable defaults with struct literal + ..Default::default().
- Convert crate-level doc comments for embedded_keys and remove repeated cfg inner attributes in macOS files.
- Misc: small API/typing tweaks (e.g. function signatures returning references) and improved readability in multiple modules.

These changes are purely refactorings/cleanups and should not alter intended behavior.

+++ src-tauri/src/macos_subtitle_overlay.rs
@@ -5,7 +5,8 @@
 //! STT / Agent とは完全に独立した機能です。
 //! また `stt-partial`（caller="live"）を監聴して発話中のリアルタイムテキストも表示します。
 
-#![cfg(target_os = "macos")]
+// The `#[cfg(target_os = "macos")]` gate lives on the `mod` declaration in
+// `lib.rs`; we don't need to repeat it as an inner attribute here.
 
 use std::cell::RefCell;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
@@ -723,7 +724,7 @@ fn show_text(app: &AppHandle, text: String, is_final: bool) {
                 ui.borrow()
                     .panel
                     .as_ref()
-                    .map(|p| p.alphaValue() as f64)
+                    .map(|p| p.alphaValue())
                     .unwrap_or(1.0)
             });
             let _ = tx.send(a);
